+++ CHANGES.md
@@ -9,7 +9,7 @@ For a full list of changes, see the [git commit log][log]
 ## v0.9.7
 
  * Drop entire ABI concept
- * Add SVCBATCH_SERVICE_LOGS variable
+ * Add SVCBATCH_SERVICE_LOGDIR variable
  * Drop console shutdown signals
 
 ## v0.9.6